requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
selenium==4.1.0
webdriver-manager==3.5.1
discord.py==2.1.1
python-dotenv>=0.19.2